fix(countdownbar): rename files in order to make more accurate to the context
