--- src/components/feedback/CountdownBar/types/AppearanceType.ts
@@ -1,3 +1,0 @@
-import { colors } from "../../../../shared/colors/colors";
-
-export type AppearanceType = keyof typeof colors.sys.actions;
