+++ src/components/feedback/CountdownBar/types/CountdownBar.Appearance.type.ts
@@ -1,0 +1,3 @@
+import { colors } from "../../../../shared/colors/colors";
+
+export type Appearance = keyof typeof colors.sys.actions;
